change long to double for task runtime2

+++ optimization-framework/src/main/java/org/optframework/core/TaskUtility.java
@@ -4,7 +4,7 @@ import org.cloudbus.cloudsim.util.workload.Job;
 import org.cloudbus.spotsim.enums.InstanceType;
 
 public class TaskUtility {
-    public static long executionTimeOnType(Job job, InstanceType type){
+    public static double executionTimeOnType(Job job, InstanceType type){
         return job.getLength() / type.getEc2units();
     }
 }
